+++ src/Events/WebDav/FileDeletedEvent.php
@@ -1,0 +1,7 @@
+<?php
+
+declare(strict_types=1);
+
+namespace N3XT0R\LaravelWebdavServer\Events\WebDav;
+
+final readonly class FileDeletedEvent extends NodeEvent {}